+++ packages/terra-profile-image/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Updated
 * updated jest snapshots
+* updated to set `fit` to always be `cover` to handle non-square images.
 
 ### Changed
 * updated package.json test scripts