+++ packages/terra-profile-image/CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ----------
+### Updated
+* updated jest snapshots
+* updated to set `fit` to always be `cover` to handle non-square images.
+
 ### Changed
 * updated package.json test scripts